Emoji Hell
A lot of times, people will go up and down the point tree. Because of
this, we get a lot of people that have the wrong emojis and what not.
Instead of trying to keep track of them, let's just reset the name each
time someone gets points, and place them in the correct "bracket", so to
speak.